added SIGUSR1 and remove files

+++ child.c
@@ -7,35 +7,45 @@
 #include <signal.h>
 #include <time.h>
 
+void signal_catcher(int);
+
 int main(int argc, char *argv[])
 {
+
+    if (sigset(SIGUSR1, signal_catcher) == SIG_ERR)
+    {
+        perror("Sigset can not set SIGUSR1");
+        exit(SIGUSR1);
+    }
     int i, childNumber;
     long t;
     FILE *ptr;
-
-    printf("im chid %d = %d\n", childNumber, getpid());
-    fflush(stdout);
-
-    childNumber = atoi(argv[0]);
-
     t = time(NULL) * childNumber;
-    printf("time from chid %d = %ld\n", childNumber, t);
-    fflush(stdout);
     srand(t);
 
-    char filename[30] = "./child";
-    strcat(filename, argv[0]);
-    strcat(filename, ".txt");
-    ptr = fopen(filename, "w+");
-
-    for (i = 0; i < 10; i++)
+    while (1)
     {
-        fprintf(ptr, "%d\n", (rand() % 100) + 1);
+        pause();
+        childNumber = atoi(argv[0]);
+        char filename[30] = "./child";
+        strcat(filename, argv[0]);
+        strcat(filename, ".txt");
+        ptr = fopen(filename, "w+");
+
+        for (i = 0; i < 10; i++)
+        {
+            fprintf(ptr, "%d\n", (rand() % 100) + 1);
+        }
+
+        close(ptr);
+        sleep(childNumber);
+        kill(getppid(), SIGINT);
     }
-
-    close(ptr);
-    sleep(childNumber);
-    kill(getppid(), SIGINT);
-
     return 0;
+}
+
+void signal_catcher(int the_sig)
+{
+    printf("SIGUSR1 received\n");
+    fflush(stdout);
 }
